Treat release candidates as build numbers.
They're still set as a different VersionType in this PR, as it
seems like the right thing to do for comparisons with beta/release.

If .parse() is called with `rc` in the string or `is_release_candidate`
is used in the constructor, then you get the `rc` variant of the string
back out again.